refactor: Make TableFunctionAnalysis into an interface
