+++ ksql-engine/src/main/java/io/confluent/ksql/planner/LogicalPlanner.java
@@ -20,6 +20,7 @@ import io.confluent.ksql.analyzer.Analysis;
 import io.confluent.ksql.analyzer.Analysis.AliasedDataSource;
 import io.confluent.ksql.analyzer.Analysis.Into;
 import io.confluent.ksql.analyzer.Analysis.JoinInfo;
+import io.confluent.ksql.analyzer.MutableTableFunctionAnalysis;
 import io.confluent.ksql.analyzer.TableFunctionAnalysis;
 import io.confluent.ksql.execution.expression.tree.ColumnReferenceExp;
 import io.confluent.ksql.execution.expression.tree.Expression;
